[ALPS05328333] ISP: P1 fixed Security Vulnerability
[Detial]
	fixed type error

MTK-Commit-Id: fa301acb61ad222c2e2c5007958b854f91645c95

CR-Id: ALPS05328333
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/cameraisp/src/mt6768/camera_isp.c
@@ -10013,7 +10013,7 @@ EXIT:
 static signed int ISP_mmap(struct file *pFile, struct vm_area_struct *pVma)
 {
 	unsigned long length = 0;
-	unsigned int pfn = 0x0;
+	unsigned long pfn = 0x0;
 
 	/*pr_info("- E.");*/
 	length = (pVma->vm_end - pVma->vm_start);
